+++ examples/with-typescript-graphql/lib/resolvers.ts
