+++ examples/with-typescript-graphql/lib/schema.ts
