+++ examples/with-typescript-graphql/lib/type-defs.graphqls
@@ -5,5 +5,5 @@ type User {
 }
 
 type Query {
-    viewer: User
+    viewer: User!
 }
