--- examples/with-typescript-graphql/with-typescript-graphql.js
@@ -1,21 +1,0 @@
-function withTypeScriptGraphQL(nextConfig = {}) {
-  return Object.assign({}, nextConfig, {
-    webpack(config, options) {
-      config.module.rules.push({
-        test: /\.graphql$/,
-        exclude: /node_modules/,
-        use: [options.defaultLoaders.babel, { loader: 'graphql-let/loader' }],
-      })
-
-      config.module.rules.push({
-        test: /\.graphqls$/,
-        exclude: /node_modules/,
-        loader: 'graphql-tag/loader',
-      })
-
-      return config
-    },
-  })
-}
-
-module.exports = withTypeScriptGraphQL
